Added all the Directories                                                                                                                                                             Added all beginning files                                                                                                                                                             Created Database server                                                                                                                                                             Copied DDL from Slack and reformated it correctly
                                                                                                                                                            Copied code from slack website and put it in files autoload.php, validateDate.php, validateUuid.php.
                                                                                                                                                            Put my git hub username and git repository in namespacing section -do have an error with git respository name think it might be because of the dash -

                                                                                                                                                            Ran composer on local machine
                                                                                                                                                            Ran composer on server

                                                                                                                                                            wrote and documented all state varibles in the class.
                                                                                                                                                            Wrote and docuented all accessor/getter method for each state variable.
                                                                                                                                                            Wrote and documented a mutator/setter method for each star variable.
                                                                                                                                                            Wrote and documented constuctor method

                                                                                                                                                            Next step i think is boss fight...

                                                                                                                                                            Added block types and fixed authorAvatarUrl acessor and mutator to match the tweet comment example. it was a string

                                                                                                                                                            Made some correction in errors with marty.
                                                                                                                                                            Errors labeling folders with class need to have capital first letter (Author.php, ValidateDate.php, ValidateUuid.php) helps json recognize file has class in them.

                                                                                                                                                            Error in syntax. namespace should have fist letter captital as well as the gitfile name.

                                                                                                                                                            also missing end json statement at the end

                                                                                                                                                            format correction that were sent to me in slack
                                                                                                                                                            wrote out the lib/author.php file

                                                                                                                                                            Corrections on running it and fixing Username so it isn't a hexidecimal

                                                                                                                                                            Successful in getting lib/author.php to run

                                                                                                                                                            Started on adding the insert, update, delete methods
                                                                                                                                                            Finished writting insert, update, and delete statements. Also did the getFooByBar method for single object and for full array. Looks like I need to look over delete statement and fix bug.
                                                                                                                                                            Fixed typo in delete statement. Fixed tab spacing so all the things under class is tab over one more.

+++ php/classes/Author.php
@@ -19,48 +19,48 @@ use Ramsey\Uuid\Uuid;
 class Author implements \JsonSerializable {
 	use ValidateUuid;
 
-/*Write and document all state variables in the class*/
-/** id for this Author is the primary key
-*@var Uuid $authorId
-*/
-private $authorId;
+	/*Write and document all state variables in the class*/
+	/** id for this Author is the primary key
+	*@var Uuid $authorId
+	*/
+	private $authorId;
 
-/** token handed out to varify that the profile is valid and not malicious.
-* @var string $authorActivationToken
- */
-private $authorActivationToken;
-/**
-* @var
-*/
-private $authorAvatarUrl;
-/**
- * email for the Author; this is a unique index
- * @var string $authorEmail
- */
-private $authorEmail;
-/**
-* hash for author password
-* @var $authorHash
-*/
-private $authorHash;
-/**
- * username for author
- * @var $authorUsername
- */
-private $authorUsername;
+	/** token handed out to varify that the profile is valid and not malicious.
+	* @var string $authorActivationToken
+	 */
+	private $authorActivationToken;
+	/**
+	* @var
+	*/
+	private $authorAvatarUrl;
+	/**
+	 * email for the Author; this is a unique index
+	 * @var string $authorEmail
+	 */
+	private $authorEmail;
+	/**
+	* hash for author password
+	* @var $authorHash
+	*/
+	private $authorHash;
+	/**
+	 * username for author
+	 * @var $authorUsername
+	 */
+	private $authorUsername;
 
-/*Write and document constructor method*/
-/**
- * Author constructor.
- * @param string|Uuid $newAuthorId if of this Author or null if a new Author
- * @param string $newAuthorActivationToken activation token to safe guard against malicious accounts
- * @param string $newAuthorAvatarUrl string containing newAuthorUrl can be null
- * @param $newAuthorEmail string containing email
- * @param $newAuthorHash string containing password hash
- * @param $newAuthorUsername string containing username
- */
-public function __construct($newAuthorId, string  $newAuthorActivationToken, $newAuthorAvatarUrl, $newAuthorEmail, $newAuthorHash, $newAuthorUsername) {
-	try {
+	/*Write and document constructor method*/
+	/**
+ 	* Author constructor.
+	 * @param string|Uuid $newAuthorId if of this Author or null if a new Author
+ 	* @param string $newAuthorActivationToken activation token to safe guard against malicious accounts
+ 	* @param string $newAuthorAvatarUrl string containing newAuthorUrl can be null
+ 	* @param $newAuthorEmail string containing email
+ 	* @param $newAuthorHash string containing password hash
+ 	* @param $newAuthorUsername string containing username
+ 	*/
+	public function __construct($newAuthorId, string  $newAuthorActivationToken, $newAuthorAvatarUrl, $newAuthorEmail, $newAuthorHash, $newAuthorUsername) {
+		try {
 		$this->setAuthorId($newAuthorId);
 		$this->setAuthorActivationToken($newAuthorActivationToken);
 		$this->setAuthorAvatarUrl($newAuthorAvatarUrl);
@@ -68,189 +68,188 @@ public function __construct($newAuthorId, string  $newAuthorActivationToken, $ne
 		$this->setAuthorHash($newAuthorHash);
 		$this->setAuthorUsername($newAuthorUsername);
 
-/*     catches errors or invalid inputs??  */
-	} catch(\InvalidArgumentException | \RangeException | \TypeError | \Exception $exception){
-		//determine what exception type was thrown
-		$exceptionType = get_class($exception);
-		throw(new $exceptionType($exception->getMessage(), 0, $exception));
+		/*     catches errors or invalid inputs??  */
+		} catch(\InvalidArgumentException | \RangeException | \TypeError | \Exception $exception){
+			//determine what exception type was thrown
+			$exceptionType = get_class($exception);
+			throw(new $exceptionType($exception->getMessage(), 0, $exception));
+		}
 	}
-}
 
-	/*Write and document an accessor/getter method for each state variable*/
-/** accessor method for authorId
-*@return Uuid of authorId
-*/
+		/*Write and document an accessor/getter method for each state variable*/
+	/** accessor method for authorId
+	*@return Uuid of authorId
+	*/
 
-public function getAuthorId() : Uuid {
-	return($this->authorId);
-}
-
-/*Write and document a mutator/setter for each state variable*/
-/**
- * mutator method for profile id
- * @param Uuid| string $newAuthorId value of new profile id
- * @throws \RangeException if $newAuthorId is not positive
- * @throws \TypeError if the profile Id is not
- */
-
-public function setAuthorId( $newAuthorId) : void {
-	try {
-			$uuid = self::validateUuid($newAuthorId);
-	} catch(\InvalidArgumentException | \RangeException | \Exception | \TypeError $exception) {
-		$exceptionType = get_class($exception);
-		throw(new $exceptionType($exception->getMessage(),0, $exception));
+	public function getAuthorId() : Uuid {
+		return($this->authorId);
 	}
-	$this->authorId = $uuid;
-}
-/* accessor for authorActivationToken */
-public function getauthorActivationToken() : string {
-	return($this-> authorActivationToken);
-}
 
-/**
- * mutator method for authorActivationToken
- * @param string $newAuthorActivationToken
- * @throws \InvalidArgumentException in the token is not a string or insecure
- * @throws \RangeException if the token is not exactly 32 characters
- * @throws \TypeError is the activation token is not a string
- */
+	/*Write and document a mutator/setter for each state variable*/
+	/**
+	 * mutator method for profile id
+	 * @param Uuid| string $newAuthorId value of new profile id
+	 * @throws \RangeException if $newAuthorId is not positive
+	 * @throws \TypeError if the profile Id is not
+	 */
 
-public function setAuthorActivationToken(string $newAuthorActivationToken) : void {
-	if($newAuthorActivationToken === null) {
-		$this->authorActivationToken = null;
-		return;
+	public function setAuthorId( $newAuthorId) : void {
+		try {
+				$uuid = self::validateUuid($newAuthorId);
+		} catch(\InvalidArgumentException | \RangeException | \Exception | \TypeError $exception) {
+			$exceptionType = get_class($exception);
+			throw(new $exceptionType($exception->getMessage(),0, $exception));
+		}
+		$this->authorId = $uuid;
 	}
-	$newAuthorActivationToken = strtolower(trim($newAuthorActivationToken));
-	if(ctype_xdigit($newAuthorActivationToken) === false) {
-		throw(new\RangeException("user activation is not valid"));
+	/* accessor for authorActivationToken */
+	public function getauthorActivationToken() : string {
+		return($this-> authorActivationToken);
 	}
-	//make sure authorActivationToken is only 32 characters
-	if(strlen($newAuthorActivationToken) !== 32) {
-		throw(new\RangeException("user activation token has to be 32"));
+
+	/**
+	 * mutator method for authorActivationToken
+	 * @param string $newAuthorActivationToken
+	 * @throws \InvalidArgumentException in the token is not a string or insecure
+	 * @throws \RangeException if the token is not exactly 32 characters
+	 * @throws \TypeError is the activation token is not a string
+	 */
+
+	public function setAuthorActivationToken(string $newAuthorActivationToken) : void {
+		if($newAuthorActivationToken === null) {
+			$this->authorActivationToken = null;
+			return;
+		}
+		$newAuthorActivationToken = strtolower(trim($newAuthorActivationToken));
+		if(ctype_xdigit($newAuthorActivationToken) === false) {
+			throw(new\RangeException("user activation is not valid"));
+		}
+		//make sure authorActivationToken is only 32 characters
+		if(strlen($newAuthorActivationToken) !== 32) {
+			throw(new\RangeException("user activation token has to be 32"));
+		}
+		$this->authorActivationToken = $newAuthorActivationToken;
 	}
-	$this->authorActivationToken = $newAuthorActivationToken;
-}
-/**
- * accessor method for authorAvatarUrl
- * @return string value of authorAvatarUrl
- */
+	/**
+	 * accessor method for authorAvatarUrl
+	 * @return string value of authorAvatarUrl
+	 */
 
-public function getAuthorAvatarUrl() : string {
-	return($this->authorAvatarUrl);
-}
-
-/**
- * mutator method for authorAvatarUrl
- * @param string $newAuthorAvatarUrl new value of author content
- * @throws \InvalidArgumentException if $newTweetContent i not a string or insecure
- * @throws \RangeException if $newTweetContent is >100 characters
- * @throws \TypeError if $newTweetContent is not a string
- */
-
-public function setAuthorAvatarUrl( string  $newAuthorAvatarUrl) : void {
-	//verify the tweet content is secure
-	$newAuthorAvatarUrl = trim($newAuthorAvatarUrl);
-	$newAuthorAvatarUrl= filter_var($newAuthorAvatarUrl, FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
-	if(empty($newAuthorAvatarUrl) === true) {
-		throw(new \InvalidArgumentException("author url is empty or insecure"));
+	public function getAuthorAvatarUrl() : string {
+		return($this->authorAvatarUrl);
 	}
-	//verify the authorAvatarUrl will fit in tbe database
-	if(strlen($newAuthorAvatarUrl) > 255){
-		throw(new\RangeException("authorAvatarUrl is too large"));
+
+	/**
+	 * mutator method for authorAvatarUrl
+	 * @param string $newAuthorAvatarUrl new value of author content
+	 * @throws \InvalidArgumentException if $newTweetContent i not a string or insecure
+	 * @throws \RangeException if $newTweetContent is >100 characters
+	 * @throws \TypeError if $newTweetContent is not a string
+	 */
+
+	public function setAuthorAvatarUrl( string  $newAuthorAvatarUrl) : void {
+		//verify the tweet content is secure
+		$newAuthorAvatarUrl = trim($newAuthorAvatarUrl);
+		$newAuthorAvatarUrl= filter_var($newAuthorAvatarUrl, FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
+		if(empty($newAuthorAvatarUrl) === true) {
+			throw(new \InvalidArgumentException("author url is empty or insecure"));
+		}
+		//verify the authorAvatarUrl will fit in tbe database
+		if(strlen($newAuthorAvatarUrl) > 255){
+			throw(new\RangeException("authorAvatarUrl is too large"));
+		}
+	//store the authorAvatarUrl
+		$this->authorAvatarUrl = $newAuthorAvatarUrl;
 	}
-//store the authorAvatarUrl
-	$this->authorAvatarUrl = $newAuthorAvatarUrl;
-}
-/**
- * accessor method for email
- *
- * @return string value of email
- */
-public function getAuthorEmail(): string {
-	return $this->authorEmail;
-}
-
-/**
- * mutator method for email
- *
- * @param string $newAuthorEmail new value of email
- * @throws \InvalidArgumentException if $newEmail is not a valid email or insecure
- * @throws \RangeException if $newEmail is > 128 characters
- * @throws \TypeError if $newEmail is not a string
- */
-
-public function setAuthorEmail(string $newAuthorEmail): void {
-	//verify the email is secure
-	$newAuthorEmail = trim($newAuthorEmail);
-	$newAuthorEmail = filter_var($newAuthorEmail, FILTER_VALIDATE_EMAIL);
-	if(empty($newAuthorEmail) === true) {
-		throw(new \InvalidArgumentException("author email is empty or insecure"));
+	/**
+	 * accessor method for email
+	 *
+	 * @return string value of email
+	 */
+	public function getAuthorEmail(): string {
+		return $this->authorEmail;
 	}
-	//verify the email will fit in the database
-	if(strlen($newAuthorEmail) > 128) {
-		throw(new \RangeException("author email is too large"));
+
+	/**
+	 * mutator method for email
+	 *
+	 * @param string $newAuthorEmail new value of email
+	 * @throws \InvalidArgumentException if $newEmail is not a valid email or insecure
+	 * @throws \RangeException if $newEmail is > 128 characters
+	 * @throws \TypeError if $newEmail is not a string
+	 */
+
+	public function setAuthorEmail(string $newAuthorEmail): void {
+		//verify the email is secure
+		$newAuthorEmail = trim($newAuthorEmail);
+		$newAuthorEmail = filter_var($newAuthorEmail, FILTER_VALIDATE_EMAIL);
+		if(empty($newAuthorEmail) === true) {
+			throw(new \InvalidArgumentException("author email is empty or insecure"));
+		}
+		//verify the email will fit in the database
+		if(strlen($newAuthorEmail) > 128) {
+			throw(new \RangeException("author email is too large"));
+		}
+		//store the email
+		$this->authorEmail = $newAuthorEmail;
 	}
-	//store the email
-	$this->authorEmail = $newAuthorEmail;
-}
 
-/** accessor method for authorHash
- *@return string value of hash
- */
-public function getAuthorHash(): string {
-	return $this->authorHash;
-}
-
-/**
- * mutator method for author hash password
- *
- * @param string $newAuthorHash
- * @throws \InvalidArgumentException if the has is not secure
- * @throws \RangeException if the hash is not 97 characters
- * @throws \TypeError if the profile hash is not a string
- */
-
-public function setAuthorHash(string $newAuthorHash): void {
-	//enforce that the has is properly formatted
-	$newAuthorHash = trim($newAuthorHash);
-	if(empty($newAuthorHash) === true) {
-		throw(new\InvalidArgumentException("author password has empty or insecure"));
+	/** accessor method for authorHash
+	 *@return string value of hash
+	 */
+	public function getAuthorHash(): string {
+		return $this->authorHash;
 	}
-	//enforce the the hash is exactly 97 characters.
-	if(strlen($newAuthorHash) !== 97) {
-		throw(new \RangeException("author hash must be 97 characters"));
+
+	/**
+	 * mutator method for author hash password
+	 *
+	 * @param string $newAuthorHash
+	 * @throws \InvalidArgumentException if the has is not secure
+	 * @throws \RangeException if the hash is not 97 characters
+	 * @throws \TypeError if the profile hash is not a string
+	 */
+
+	public function setAuthorHash(string $newAuthorHash): void {
+		//enforce that the has is properly formatted
+		$newAuthorHash = trim($newAuthorHash);
+		if(empty($newAuthorHash) === true) {
+			throw(new\InvalidArgumentException("author password has empty or insecure"));
+		}
+		//enforce the the hash is exactly 97 characters.
+		if(strlen($newAuthorHash) !== 97) {
+			throw(new \RangeException("author hash must be 97 characters"));
+		}
+		//store the hash
+		$this->authorHash = $newAuthorHash;
+
 	}
-	//store the hash
-	$this->authorHash = $newAuthorHash;
-
-}
-/**
- * accessor method for authorUsername
- */
-public function getAuthorUsername(): string {
-	return ($this->authorUsername);
-}
-
-/** mutator method for authorUsername
- * @param string $newAuthorUsername
- * @throws \RangeException if the toke is not 32 characters
- * @throws \TypeError if the activation token is not a string
- */
-
-
-public function setAuthorUsername (string $newAuthorUsername): void {
-	$newAuthorUsername = trim($newAuthorUsername);
-	$newAuthorUsername = filter_var($newAuthorUsername, FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
-		if(empty($newAuthorUsername) === true) {
-			throw(new\InvalidArgumentException("username is not valid"));
+	/**
+	 * accessor method for authorUsername
+	 */
+	public function getAuthorUsername(): string {
+		return ($this->authorUsername);
 	}
-		if(strlen($newAuthorUsername) >32) {
-			throw(new \RangeException("author username has to be 32 characters"));
+
+	/** mutator method for authorUsername
+	 * @param string $newAuthorUsername
+	 * @throws \RangeException if the toke is not 32 characters
+	 * @throws \TypeError if the activation token is not a string
+	 */
+
+
+	public function setAuthorUsername (string $newAuthorUsername): void {
+		$newAuthorUsername = trim($newAuthorUsername);
+		$newAuthorUsername = filter_var($newAuthorUsername, FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
+			if(empty($newAuthorUsername) === true) {
+				throw(new\InvalidArgumentException("username is not valid"));
+		}
+			if(strlen($newAuthorUsername) >32) {
+				throw(new \RangeException("author username has to be 32 characters"));
+		}
+		$this->authorUsername = $newAuthorUsername;
+
 	}
-	$this->authorUsername = $newAuthorUsername;
-
-}
-
 
 	//Insert statement method
 
@@ -327,14 +326,13 @@ public function setAuthorUsername (string $newAuthorUsername): void {
 	public function delete(\PDO $pdo) : void {
 
 		//create query template
-		$query = "DELETE FROM author
-						WHERE authorId = :authorId";
+		$query = "DELETE FROM author WHERE authorId = :authorId";
 
 		//prepare a statement object using the SQL so pdo knows what to do.
 		$statement = $pdo->prepare($query);
 
 		//bind the member variables to the place holder in the template
-		$parameter = ["authorId" => $this->authorId->getBytes()];
+		$parameters = ["authorId" => $this->authorId->getBytes()];
 
 		//now execute the statement on the database.
 		$statement->execute($parameters);
@@ -412,14 +410,14 @@ public function setAuthorUsername (string $newAuthorUsername): void {
 	 *
 	 * @return array resulting state variables to serialize
 	 */
-public function jsonSerialize() : array {
-	$fields = get_object_vars($this);
+	public function jsonSerialize() : array {
+		$fields = get_object_vars($this);
 
-	$fields["authorId"] = $this->authorId->toString();
-	unset($fields["authorActivationToken"]);
-	unset($fields["authorAvatarUrl"]);
-	unset($fields["authorHash"]);
-	unset($fields["authorUsername"]);
-	return($fields);
-	}
+		$fields["authorId"] = $this->authorId->toString();
+		unset($fields["authorActivationToken"]);
+		unset($fields["authorAvatarUrl"]);
+		unset($fields["authorHash"]);
+		unset($fields["authorUsername"]);
+		return($fields);
+		}
 }
